Implement fit_sqw. Required various additional routines to be written, and also requires that multifit is added to libisis or mgenie.
multifit is a generalisation that allows simultaneous fitting to many objects - not just sqw or d1d, d2d,... objects. Multifit is complete as it stands. It needs one or two additional features to be added to enable multifitting to be implemented directly in Horace. This will be added with the first update to the official release.

refs #63

git-svn-id: https://svn.isis.rl.ac.uk/Horace/branches/sqw@196 1cb955da-4463-431d-88dd-2af2f5550de3

Former-commit-id: 1104076b2c5bb5673ed00e65f682c09d6a0345da