Implement calculation of a dispersion relation as a dnd object of dimension equl to the input sqw or dnd object(s), if energy is not one of the display axes of the input object, or dimensionality one less than the input if energy is one of the axes. Overplotting in one dimension is already implemented.
Also: to avoid errors when cutting, now require that the energy offset is zero in uoffset fields and in the 'proj' input to cut_sqw. The cut algorithm does not have this possiblity implemented - and who would want it anyway? Constructors for d0d,d1d,...d4d altered to ensure consistency.

refs #14 #24

git-svn-id: https://svn.isis.rl.ac.uk/Horace/branches/sqw@202 1cb955da-4463-431d-88dd-2af2f5550de3

Former-commit-id: 4c82b456e901bfc43855ed3af017b50e6d2a08b4