refs #76 tortoseHoracehook.bat changed to work with perl installed with Matlab the hook updates horace_version.m in accordance with the latest SVN version.
The hook needs proper settings for MATLABDIR environment variable (necessary for Libisis development under windows) and expect the Libisis development edition to be installed in the same subfolder as Horace. 
If the later is not the case, the svn_matlab_commit_hook.pl  has to be make available for for Horace or the tortoseHoracehook.bat script modified to make this script available. 


git-svn-id: https://svn.isis.rl.ac.uk/Horace/trunk@336 1cb955da-4463-431d-88dd-2af2f5550de3

Former-commit-id: 724757cfc5f07e1d6c4e8c2bce014ef3e7d3c17a