Fix linker error when building combine_sqw
A linker error was raised when compiling combine_sqw due to using the
function 'utIsInterupping'. The function is contained within the `libut`
library that ships with Matlab.

Add a horace_FindMatlab script, which calls `find_package(Matlab)` and
defines a `Matlab_UT_LIBRARY` variable that points to `libut`. Link the
library to the combine_sqw target.