Ensure MX library is found in horace_FindMatlab
On Unix (tested on Ubuntu 18) CMake's built-in FindMatlab script does
not export the path to Matlab's MX library. This library is needed for
building tests that use the mex library.

Add an `if` statement into `horace_FindMatlab` to check the variable
pointing to the MX library is defined. If it is not, find the library.


Former-commit-id: f0addfdde2469b4fa37c4f7f42ae69805fcefc8a