Add optional CMake flag to build Fortran mex
The option will allow the building of only the C++ mex libraries. If you
don't have a Fortran compiler then you can pass this flag and your build
will not fail.

This option is OFF by default, i.e. Fortran libraries won't be built.
The reason for this is due to CMake not being able to use multiple build
tool-chains within the same project. I.e. on Windows, if we wanted to
build the C++ using MSVC, we cannot then use MinGW's gfortran to build
the Fortran (I haven't tried the Intel Fortran compiler). This is not a
problem on Linux if you're using the GNU tool-chain since it provides a
C, C++ and Fortran compiler.

There are workarounds for using a different toolchain (building the
Fortran as a separate CMake project, see
`cmake_add_fortran_subdirectory`
https://cmake.org/cmake/help/latest/module/CMakeAddFortranSubdirectory.html)
However the rewards for doing this do not meet the effort required.