Update Jenkins build scripts for new plugin
A new plugin (Generic webhook trigger) now takes care of receiving
GitHub webhooks, parsing their response JSON and triggering a build if
required. This means we no longer need to check any branch out within
the Jenkinsfile script.

This ultimately make the Jenkinsfile more general, as we assume the
correct branch is checked out when it is called.


Former-commit-id: b1000aa6b51e1ec94de9206d16b4c8d1a428808f