Notify GitHub of Jenkinsfile build status
When a build's pipeline starts the pending notification will be sent to
the PR's head commit. If the build succeeds or fails, the relevant
status will be sent. This notification is sent using `curl`.

Note that the 'pending' status may take a minute or two to come through,
since Jenkins checks out the code before starting the pipeline (which is
where we can call the `curl` command.