Use JOB_BASE_NAME not BASE_NAME in GitHub statuses
When Jenkins sends GitHub status updates it sends a `context` which was
previously the Jenkins defined JOB_NAME variable. This would have been
something like `PACE-neutrons/Scientific-Linux`. To neaten the view up
on GitHub we now use the JOB_BASE_NAME variable as the context. The
displayed name in the previous example will now be `Scientic-Linux`.


Former-commit-id: a3434869b828f2c6295aaab3024cf62175682521