Move Jenkins pending status into node block
Move the code that sends the "pending" job status to GitHub into a node
block. Previously the notification was sent with the main pipeline, but
this meant there was a large delay as the repo needed to be checked out
and a builder assigned before the status was sent.

Moving it into a node block may reduce the time waiting for the
notificatino to be sent.


Former-commit-id: 8103f3dda834c9fc6962184e717004fe688259e9