Update Jenkins SMG "Pipeline" config image
The refspec that was prevously checked out did not include master, so
when the "merge with master" step was being carried out it was being
done against an old version of master. This old version was apparently
still hanging around in the workspace from a previous build.

Update the refspec in the image to match the refspec now in the Jenkins
GUI (i.e. fetch origin/master too!) so an up-to-date master is used.


Former-commit-id: 6b771652db5c66f9c08987eacecd55db7c0f1bef