Add CMake herbert_add_mex function
This function replaces the uses of `matlab_add_mex`. The reason for
this is that we want to copy all the built mex files to the
<herbert_root>/DLL directory so they can be picked up by Matlab. The
function passes all its arguments to the `matlab_add_mex` function and
then adds a post-build command to the target which copies it into the
desired directory.