Fix bug with Jenkins not finding package archive
The glob used to find the archive package did not work, I'm not sure
what globbing rules archiveArtifacts uses because the glob worked
locally using `ls`. The rule used now is sufficient and should work. We
just need to ensure that no other files are outputted to the top level
of the build directory that start with "Herbert-"