Use double-quotes for shell tar extract command
The shell command to extract the tar archive that was copied from
Herbert uses a wildcard since we don't know what it will be called.
Using single-quoted strings for the shell command, Jenkins executed the
command as
"tar --wildcards --strip-components=1 -xf 'Herbert-*' '*/Herbert'" with
quotes around the arguments proceeding the `-xf` flag. This causes an
error as the wildcard won't work with the quotes there. Hopefully using
double-quoted strings will prevent this.