Add target to copyArtifact command in Jenkinsfile
After adding an ls command after the copyArtifact call, it appears that
the artifact was not being copied, despite what Jenkins was saying. Try
adding a target directory to see if it's copied.

Try adding .toString() to the end of the bash command string to avoid
Jenkins putting single quotes around the command line arguments.


Former-commit-id: e34eee53319d0ae28a1b20e387a99750a5e5d226