Use `find -name` to find Herbert tar in Jenkins workspace
The copyArtifacts call was finding the `build/Herbert-*.tar.gz` file and
copying it into the `build` directory in the current workspace - not the
working directory.

Using `find -name` is a more robust method of getting the correct path
to the .tar. We still need to use `ls` or `find` since the exact name of
the tar is not known - it contains a git sha.