Write a FindHerbert.cmake script
To run the Matlab tests in CTest the path to the `herbert_init` script
is required.

At configure time, CMake will look for `herbert_init.m` in a list of
directories given in the FindHerbert script. If Herbert is not found,
the build will NOT fail, however a warning will be shown that Herbert
could not be found.