Pacakge _test directory with Herbert
The reason for this comes mostly from Jenkins+Horace. Horace pulls the
Herbert build package when it runs its tests, however the testing
framework is in Herbert! Hence the test framework needs to be present in
Herbert - at least for the time being.

Ideally we would not package the _test directory at all and Horace's
tests should all be written using the new Matlab test framework. This is
probably not possible, however we should find a way to minimise the size
of the Herbert package we ship to users, but still allowing Horace to
perform its full set of tests etc.