Rethrow error caught in horace_on.m.template
An error is caught within horace_on if an error is thrown when
attempting to run herbert_on. This was swallowing the error message so
it couldn't be worked out why herbert_on was failing. Displaying the
previous error message and then rethrowing the error gives more detail
about the source of any problems.