Add Windows commands in Clone-Herbert Jenkins stage
Adds the Windows command to extract the archive pulled from Herbert.
Powershell's `Expand-Archive` is a bit limited in its features, so
there's a little hackery - we extract the entire zip, pull out the
folders we need (i.e. the `Herbert` folder) and bin the rest.

We don't do much extra work doing it this way, as the only files in the
archive that don't live in `Herbert` are the licence, readme and a
couple of Matlab scripts.