Check for Herbert existence before copying artifact
On Jenkins (on Windows), after the Herbert artifcat .zip file is copied
the "Herbert" directory is copied into the workspace directory. However,
if a Herbert directory already exists (e.g. if a previous build is not
cleaned up properly), Windows will not overwrite it and an error will
be raised. So check for the existence of a Herbert file/directory and
delete it if required.

Also delete the extraction folder within the system temp directory.


Former-commit-id: 0e0f83c9ae2e5931bdaa6e8cd091e870d216a239