Re-enable checks for C++ API in Matlab CMake
The checks for whether Matlab_ENGINE_LIBRARY and
Matlab_DATAARRAY_LIBRARY were found in FindMatlab were disabled. This
was because errors were being thrown when building against R2015b. This
was actually due to an error with CMake's FindMatlab script not
recalculating which version it has found after the Matlab_ROOT_DIR is
changed. It still thought it was pointing at version 9.5 (which has the
C++ API) when it was actually pointing at 8.5. This meant it was
checking for libraries that did not exist in v8.5 and throwing an error.