Package Unix MPICH within repo
The version of MPICH Matlab uses (v1.4.1p1 - information I got from
calling `strings` on Matlab's MPICH .so libraries and grepping for a
version number) is not easily installed - I had to build it from source.
MPICH is also not available on ANVIL's sl7 machines, the easiest thing
to do was to package MPICH and build against is statically.

Note you may still require "pthread" to be installed on the system you
run Herbert on.