Set TMP environment variable when running CTest
A test was failing on the Windows server because a temporary file
written to "User\AppData\Local\Temp" was left over from a
previous/parallel build. Setting the TMP variable to somewhere within
the Jenkins workspace (in the CMake build directory) means we avoid
these collisions.

The problem file was "test_gen_runfiles_spe_1.nxspe"


Former-commit-id: 45e31fcccb898ddcf5fc6da2c7a31dc4eadb03f2