Rename herbert_Version.cmake -> PACE_Version.cmake
The script has now been made general, so that it can also be used by
Horace. Instead of a fixed prefix (e.g. Herbert) the script now uses the
name of the project to prefix its variable names. The main output
variable is now "${PROJECT_NAME}_VERSION" - which will evaluate to
"Herbert_VERSION" for Herbert and "Horace_VERSION" for Horace.