Check for presence of herbert_init in FindHerbert
I ran into an issue where, when running CMake, if I passed it a path
to Herbert_ROOT, but the path was one level too high CMake would think
it had found Herbert, in fact it finds all the files it needs. However
because Herbert_ROOT was not the directory containing herbert_init.m,
the wrong path is written to horace_on.m and the tests would not run.

The only way to run into this problem is to pass the path in yourself,
otherwise CMake explicitly looks for herbert_init.m. So throwing an
error telling the user the path doesn't contain herbert_init.m should
always avoid this bug in the future.


Former-commit-id: 678850436e6ca324ea2cfbd76b6279d6af224804