Move find_path call into if block in FindHerbert
Do not call find_path(Herbert_ROOT) in CMake if Herbert_ROOT is defined
but herbert_init.m does not exist within it. This change makes no
functional difference (find_path will do nothing if Herbert_ROOT is
already defined), but makes the intention of the `if` block more
explicit.


Former-commit-id: 9f45aa84c9a93922ea220a43de9670751074450e