Make sure Herbert_ROOT is always in CMake GUI
Previously, an error was raised straight away if Herbert_ROOT was passed
on the command line but the directory did not contain `herbert_init.m`,
meaning no `find_path(Herbert_ROOT)` call was made. This meant that
there was no Herbert_ROOT variable in the CMake GUI.

To ensure the variable is always present and editable in the GUI, make
sure we call `find_path(Herbert_ROOT)` before we raise an error.


Former-commit-id: d954c8f40965d614f2c79510347212bddeb47698