Use Jenkins `unsuccessful` block to post statuses
Previously we were using the "failure" stage to post unsuccessful build,
statuses to GitHub, however this block is not entered if the build is
manually aborted. This meant that aborted jobs were stuck with the
"build running" status on GitHub.

The `unsuccessful` block is entered if the build has any status other
then `success`.


Former-commit-id: 06b8e201a8bf1752c496ae29535cdf9125d97e30