Set TMP environment variable when running CTest
The value of the TMP environment variable is what's returned by Matlab's
`tempdir` function. TMP has been set to within the CMake build
directory, this ensures the tests are not writing files to the system's
temporary directory. If two processes are running the same test (e.g. on
a build server), then the processes can run into IO race conditions.