Use Matlab ASCII reader for gen_sqw_cylinder test
The Matlab and mex readers gave slightly different answers due to
floating point errors, this was causing inconsistent binning of data in
the test. This meant the test only passed with the Matlab version,
since that was what was used to generate the reference file.

An additional test will be added to Herbert to verify the ASCII readers
do not differ by more than 1e-5 (the values in the .par files are only
specified to 4 decimal places).