Fix and re-enable combine_sqw.test tests
The issue was that the sqw file was not being read correctly. The
`ifstream::read` call was reading pixels in steps of 8 bytes, but
allocating them to an array holding floats (size 4) - so the array was
only being half filled.

I'm not sure why this was working on GCC and MSVC 2019 - maybe the
compiler was doing some clever narrowing casts that meant the values
were converted to floats.


Former-commit-id: 500afc27ae81a87371983118a5b05effb35e931d