Pass raw pixel data to accumulate_cut mex
Remove skipping of mex test if use_mex is false. I don't see any value
in skipping the test in use_mex is off, the test is specifically testing
that mex file, if not testing mex, just don't run the test. We're more
likely to accidentally skip the test when we don't mean to


Former-commit-id: 8347bd8e6104fa7b37a1dd421461b2afeb0fe668