Make PixelData a copyable handle class
This allows us to set data on PixelData class instances without
returning a copy of an object. For example, PixelData that is
file-backed should not load any data into memory on construction.
Instead, data is loaded when it is requested and held in memory for
optimization reasons. It is impossible to write a getter that returns
only data and sets data loaded from a file on the class instance. A
handle class can do this.

A handle class allows us to:
  a) lazy load data on request
  b) cache data on class instances when getting data
  c) load pages when given data ranges are passed to a getter