Add 'advance' method to PixelData class
This function loads the next page of data from a file backed PixelData
instance. This method, combined with PixelData.has_more() allows users
to loop over all the data within a file-backed PixelData instance.

PixelData.advance does nothing if there is no file backing the PixelData
object. It throws an error if attempting to advance past the last page
of data in the SQW file


Former-commit-id: 24ed91d5a2f8d031a93075fb7df4089b7be89c99