Add copy to SQW to deal with handle class PixelData
A simple copy-on-assignment no longer works for an SQW object since
PixelData is now a handle class. If copying on assignment, both SQW
objects will point to the same pixel data, meaning a change in one will
result in a change to the other.

The new copy method explicitly copies the handle class PixelData, so
each object has its own copy of the pixel data