Set PixelData.page_size on construction
After construction, a PixelData object used to have page_size = 0, this
was because there were no pixels in memory until a getter was called.
This has been a bit annoying when writing loops that use the page size.

As far as the API is concerned, having no pixels in memory should
not make a difference.