Throw 'MATLAB:nonExistentField' for unknown sqw field
No error ID was previously given to the error being thrown. This allows
us to use a try-catch to check if a field exists on a dnd/sqw object.
The usual "isprop" and "isfield" functions do not work on dnd/sqw
objects - I'm not clear on the reason for this, but is likely to do with
the fact they are old-style classes


Former-commit-id: 8ed03a065b847f2d152f393054bf1f416db6132a