Remove upper-bound on pixel page size config option
I think it is best left up to the user what to set this too. We could
set the upper-bound to be the amount of RAM available on the machine,
but this may still exceed the amount of free memory. We may set the
value to the memory().MaxPossibleArrayBytes, but this is dynamic and the
user may end up with more or less memory than this when they construct a
PixelData object - the function is also not available on Linux or OSX.

Letting the user set it also gives them the opportunity to
effectively disable paging by setting the option to be more than the
amount of memory on the system