Add ability to build PR label branch in Jenkins
Adds the ability to build against a Herbert branch specified in a
pull-request label.

When a PR build is triggered, the webhook's JSON is parsed for the pull
request's labels. If the one of the labels follows the pattern
"Herbert_<some_git_branch>" then "some_git_branch" will be pulled from
Herebrt and used in the build.


Former-commit-id: 6b11f1e61d3c013ed2f3e04c4754a0abd1254a18