Add ability to build PR label branch in Jenkins
Adds the ability to build against a Herbert branch specified in a
pull-request label.

When a PR build is triggered, the webhook's JSON is parsed for the pull
request's labels. If the one of the labels follows the pattern
"Herbert_<some_git_branch>" then "some_git_branch" will be pulled from
Herebrt and used in the build.


Former-commit-id: ff6911eb502a706c29550908ecb1140076a34958