Print the RNG state in test_tobyfit_refine_crystal
The test is failing every now and then, although the RNG should
guarantee that the same random numbers are generated every time it's
run. So we'd expect it to always pass, or always fail.

I've added some print statements that show the state of the RNG at key
points in the test.