Add test function set_config_option
Use this function in test_PixelData_operations tests.

Saving/restoring config options is a little bit slow and was being done
in setUp and tearDown in test_PixelData_operations. This function
provides a clean way to temporarily set a config option which doesn't
require the option to be set for every test in the class


Former-commit-id: 4f8b2e2840ea3634b6a2d84781f261c438db73a1