Do not default to Visual Studio 2017 in build.ps1
Allow CMake to pick the default compiler, this will be the most
up-to-date Visual Studio on the system. The script still allows one to
specify a specific Visual Studio version

This was causing issues when attempting to use the build script to
re-build targets originally build with Visual Studio 2019. The script
would default to attempting to rebuild with 2017, which would upset
CMake and cause an error