Handle out of range logical indices in get_pixels
The behaviour is intended to mimic the behaviour of a Matlab array being
indexed with a logical array: if the logical array contains any "true"
values outside the target array range, throw an error. If the logical
array only contains "false" values outside the target array's range,
ignore the values.


Former-commit-id: d66792623932d60130b7753c6bffb425c7eaab66