Fix in bug PixelData when using get_pixels
When config option pixel_page_size is small, when creating a new
non-paged PixelData object from a paged one through .get_pixels, the new
unpaged object will think that it is file-backed. This caused errors
when trying to access data, as it was trying to load from file when
there is no file backing the object

This was caused by the calculate_page_size_ method which was using the
config option to calculate how many pages it expected the data to be
separated into. If the amount of pixels in memory exceeded the number of
pixels that could fit in the pixel_page_size memory allocation, the
is_file_backed_ method would return true.

Change calculate_page_size_ to return the maximum of the current page
size and the page size deduced by calculcate_page_size. This addresses
the case where the PixelData object is fully in memory and contains more
pixels than would fit in a single page with memory allocation
'pixel_page_size'.