Remove pixel range arguments from put_pix
These arguments must be given at a very low-level in the API and are not
used anywhere within Horace. The functionality is not tested anywhere
and during manual testing it was found that using the pixel range
arguments ended up corrupting the file anyway.

The functionality has been removed. It can be restored, fixed and tested
at a later date if it's required.


Former-commit-id: 6f0c336aaefef3caaa42162ec5245fcc8632e83d