Fix PixelData.get_pixels bug when cache is dirty
The PixelData.get_pixels function was only attempting to load pixels
from the backing sqw file or temporary files and was therefore ignoring
any changes to the currently cached data.

Assign currently cached data first to fix this. This also has the
benefit of not having to re-load any cached data from file.


Former-commit-id: 95d030ff6136647ab14309f962150b85f5216e30