Add PixelData.get_pixels_in_ranges method
This allows one to load pixels that exist in ranges of indices. This
method has been added to improve the performance of cut. The cut
routine would work out what ranges of pixels belong to the bins that
it's interested in, we were then converting those ranges to indices to
pass to PixelData.get_pixels, then back to ranges again when loading
from file.


Former-commit-id: 32a5bd66f2661f8b2f062d4395ccfc91975145a5