Fix error when calling cut_dnd on file path
The call to cut_dnd was calling `dnd.cut` which then called
`sqw(file_path)`, which failed because the file was not a valid SQW file
(it was a .d2d file).

This also fixes errors not being thrown when dnd objects were passed to
cut_sqw.

When the DnDBase class is implemented we should be able to get rid of
the code that casts dnd objects to sqw and back again.


Former-commit-id: 6b871d0cb31705f7b087d7b7b142ea77aee4816d