Relax tolerance in test_rebin test case
Changes to cut seemed to cause the test to fail, with a difference
between values of ~1.33e-6, where the tolerance was 1e-6. Note that this
only occurred when mex was on, using the Matlab implementation of
accumulate_cut_c, the test passed OK.

It was agreed the difference was negligible and the tolerance could be
increased.


Former-commit-id: 5e1a6085717b61b6aa58d21fcc59427d4636f693