Recalculate pix_combine_info.pix_range on accumulate
The pix_range attribute was not being updated on pix_combine_info
objects when accumulate_pix_to_file_ was called. This was causing the
pix_range to always be the default (empty), which, in turn, caused an
error when faccess_sqw_v3_3 attempted to figure out if it needed to
recalculate the pix range or not.