Fix PixelData tests when parallel dir not temp dir
The place where PixelData objects hold their temporary files changed
from tempdir() to get(parallel_config, 'working_directory') in 1497897.
This went undetected at first because the two directories are often the
same, but many Windows builds are now failing because some PixelData
tests are still looking for the temporary files in tempdir().

The tests have been updated to look for the temp files in the correct
place.