Replace uses of CMAKE_SOURCE_DIR with Horace_ROOT
It's best practice to avoid using CMAKE_SOURCE_DIR where possible. It's
becoming more common for CMake projects to import other CMake projects
as part of their build (similar to what we do with GTest). Because of
this, if another project imports the Horace CMake build,
CMAKE_SOURCE_DIR may refer to the parent project and not Horace's root.

Horace_ROOT is defined in the top-level CMakeLists file and should be
preferred to CMAKE_SOURCE_DIR.


Former-commit-id: 5165b6504c02f47769344f672d8430f97b949089