Add nosave option to PixelData.advance
This allows the user of the class to advance to the next page of data
whilst discarding any cached changes. This can be useful when you want
to make a change to a PixelData object, but preserve the original state
of the object; with the 'nosave' option you can avoid copying the
potentially very large object.


Former-commit-id: 2bf6cdab4774ce4895a1b11243d40dbfa2547a22