Fix put_sqw error with args -nopix & -reserve
The put_sqw function was recently updated to return early when using the
-nopix flag, so that pixels could be written separately. However,
returning early was causing a problem when the -reserve flag was also
given.

In the case of -reserve zeros are written where pixels would normally be
written, so there's no problem writing the footers after.

This commit updates put_sqw to not return before writing footers if the
-reserve flag is passed.