Add default empty value to Jenkinsfile params
This is a fix for errors in the Jenkins master builds which are
triggered on a timer. Since these builds are triggered automatically,
parameters with no default value are set to `null` which is not allowed
by Jenkins.

Giving all parameter values a default of an empty string allows the
build to proceed, and still gives us a false-y value that allows us to
override the empty value in `get_param`.