Add horace_install script
This script will install Horace from a release package.

At the moment, the script cannot install Horace from repo checkout,
however the function is written such that it should be possible to do
so. To do it, you could add an argument to the script that adds some
search paths when it looks for files.


Former-commit-id: 5f20e20cb9dc712a255702b7aa51645c0f86bff0